Every Function works with the 2D vector containing pairs that contain a player pointer and an int.
Some bug fixing with the new changes, specifically within CheckIfBothOwner() to compare things correctly.

CheckIfExistingPath() and CheckPathHelper() seem to work better here not printing out incorrect stuff.

CheckPathHelper() now looks at the .second of the board to check for paths instead of the .first, SHOULD NOT be passed by reference, shizzy memory problem both in leakage and in expanded. Copying the board may be necessary but the leakage is not, gotta delete our heap memory.

Basically works...? Need to bug hunt but each phase works.