De-allocation has been added for the end of the game. We are most likely not free of memory leaks though as CheckPathHelper() is not being passed in by reference. Winning and loosing works, but just un-flavorful, might need to be spiced up later.
There are still some problems with when to print the board to make it clearer of how many troops are at each province. Minor though.

This fork seems to be finished, the next one will be reworking CheckPathHelper() so that it works with being able to be passed by reference preventing leaks.